pyproject.toml
[tool.poetry]

name = "clangen"
version = "0.7.9"
description = "Fan-edit of the warrior cat clangen game"
license = "MPL-2.0"
authors = [
    "Sablesteel <noemail@null.com>"
]
readme = "README.md"
homepage = "https://clangen.io"
repository = "https://github.com/ClanGenOfficial/clangen"
keywords = ["warriorcats", "fanedit", "game", "pygame"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.7.2,<3.13"
pygame-ce = "^2.2.1"
ujson = "^5.7.0"
pygame-gui = "0.6.9"
platformdirs = "^3.3.0"
pypresence = {version = "^4.2.1", optional = true}
pgpy = "^0.6.0"
requests = "^2.28.2"
strenum = "^0.4.10"
pyinstaller = "5.13.2"


[tool.poetry.group.test]
optional = true
[tool.poetry.group.test.dependencies]
pylint = "^2.17.3"

[tool.poetry.group.build]
optional = true
[tool.poetry.group.build.dependencies]
pyinstaller = "^5.13.2"
pillow = "^9.5.0"
setuptools = "^67.7.2"
wheel = "^0.40.0"

[tool.poetry.extras]
discord = ["pypresence"]
